LIBITD-1447. Initial working implementation using CASCADE DELETES
Simplified the code by analyzing the schema to determine which tables
have CASCADE DELETES, meaning that a deletion of a row in a parent
table will automatically result in the deletion of the related rows
in a child table. This simplifies the number of "purge" queries that
must be run.

Also added in "cleanup" queries to clear orphaned records (records
that should have an entry in a parent table, but don't). These records
were discovered by deleting all entries in the "redcap_projects" and
"redcap_user_information" tables, and then seeing what records were
left in their child tables.

https://issues.umd.edu/browse/LIBITD-1447