LIBITD-1447. Updated parameters in env_example
The "SUMMARY_OUTPUT_FILE" and "RELATED_ITEM_IDS_OUTPUT_FILE" are no
longer needed and so were removed.

Moved the "CLEANUP_OUTPUT_FILE" parameter to be above the
"PURGE_QUERIES_OUTPUT_FILE", to have consistent ordering (i.e., the
order in which the files would be run against the database).

https://issues.umd.edu/browse/LIBITD-1447