LIBITD-1447. Removed code clearing null project id/user name rows
Apparently in some tables a null project id or user name is taken
as a "global" or "system" entry.

When removing the null ui_ids from the "redcap_messages_recipients" and
"redcap_surveys_themes", a notification is displayed in the
"Control Center" of the RedCap application, indicating that these rows
are missing and should be recreated.

Therefore, it seems inadvisable to remove "null" project ids or
user names from any of the tables.

https://issues.umd.edu/browse/LIBITD-1447